requirements.txt
streamlit==1.51.0
crewai==1.7.2
crewai-tools==1.7.2
python-dotenv==1.1.1
groq==0.37.1
sentence-transformers==5.2.3
